Cleaned up GUIs
All GUIs are loaded when switched to.
When closing out the application, it makes sure you want to exit with a
message box.  If no or X, you will stay in the game, if yes, it will
exit.  We can put more code in that exit to clean up the game before
they exit like leaving a table, saving some stats, and logging them out.